requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
backports-abc==0.5
beautifulsoup4==4.6.3
bleach==3.1.4
cachetools==4.0.0
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
coverage==4.5.1
cryptography==2.8
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
docopt==0.6.2
docutils==0.14
docx2pdf==0.1.5
entrypoints==0.2.3
et-xmlfile==1.0.1
EZGmail==0.0.8
funcsigs==1.0.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.17.0
hypothesis==3.66.2
idna==2.7
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.5.0
MarkupSafe==1.0
mistune==0.8.3
mock==2.0.0
more-itertools==4.3.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.8
numpy==1.18.1
oauth2client==4.1.3
openpyxl==2.5.6
packaging==17.1
pandocfilters==1.4.1
parso==0.3.1
path.py==11.0.1
pathtools==0.1.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
pkgconfig==1.4.0
pluggy==0.7.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyrsistent==0.15.7
PySocks==1.6.6
pytest==3.7.4
pytest-watch==4.2.0
python-dateutil==2.7.3
python-docx==0.8.10
pytz==2018.5
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.4.1
requests==2.20.0
rsa==4.0
Send2Trash==1.4.2
setuptools-scm==3.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.3.0
terminado==0.8.1
testpath==0.3
tornado==5.1
tqdm==4.43.0
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.24.2
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
xlrd==0.9.4
xlwt==1.3.0
